fix: move handoff state to shared services
